+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # Changelog
 
+## 4.0.8 - 2023-12-08
+
+### Fixed
+- Fix LinkedIn not switching from an image post to an article post, when containing both a link and image.
+- Fix LinkedIn not respecting custom Guzzle config when fetching images when posts contain an image.
+
+## 4.0.7 - 2023-10-25
+
+### Fixed
+- Implement `Element::trackChanges()` for Blitz compatibility.
+
+## 4.0.6 - 2023-10-06
+> {note} If you are using LinkedIn, your LinkedIn app will need to include new products. Refer to the [docs](https://verbb.io/craft-plugins/social-poster/docs/providers/linkedin).
+
+### Changed
+- Change LinkedIn from using deprecated “Shares API”.
+
+## 4.0.5 - 2023-10-05
+
+### Changed
+- Change LinkedIn to use new OpenID Connect API.
+
 ## 4.0.4 - 2023-07-11
 
 ### Added
@@ -127,12 +149,12 @@
 ## 2.1.1 - 2020-04-13
 
 ### Added
-- Add support Linked.in for company pages. See [docs](https://verbb.io/craft-plugins/social-poster/docs/providers/linked-in).
-- Updated Facebook and Linked.in provider docs with more detailed instructions.
+- Add support LinkedIn for company pages. See [docs](https://verbb.io/craft-plugins/social-poster/docs/providers/linked-in).
+- Updated Facebook and LinkedIn provider docs with more detailed instructions.
 
 ### Changed
-- Remove `r_basicprofile` and `w_share` permissions for Linked.in. These are no longer valid.
-- Remove `visibility` settings for Linked.in. No longer valid.
+- Remove `r_basicprofile` and `w_share` permissions for LinkedIn. These are no longer valid.
+- Remove `visibility` settings for LinkedIn. No longer valid.
 
 ### Fixed
 - Fix checking for valid asset field and asset.
@@ -147,7 +169,7 @@
 ## 2.0.10 - 2020-01-10
 
 ### Fixed
-- Fix posting to Linked.in
+- Fix posting to LinkedIn
 
 ## 2.0.9 - 2020-01-10
 
@@ -165,7 +187,7 @@
 ## 2.0.8 - 2019-04-24
 
 ### Changed
-- Remove `r_liteprofile` for Linked.in.
+- Remove `r_liteprofile` for LinkedIn.
 
 ### Fixed
 - Fix error when trying to post to Twitter.
@@ -266,10 +288,10 @@
 ## 1.1.0 - 2016-07-03
 
 ### Added
-- Added support for Linked.in as a provider.
+- Added support for LinkedIn as a provider.
 - Added support for Facebook Pages, along with Personal timelines.
-- Added Title field to set the post title for Facebook and Linked.in posts.
-- Added URL field to allow specific URLs to be used with Facebook and Linked.in posts.
+- Added Title field to set the post title for Facebook and LinkedIn posts.
+- Added URL field to allow specific URLs to be used with Facebook and LinkedIn posts.
 - Added new setting for auto-post. Handy if you want the widget to appear, but not be default to on.
 
 ### Changed